Split chunkgen into separate crates
Motivation: The client and the server are growing more and more different, while sharing many common types. Their behavior differences means they should be split up, but they should still be able to access the core types `chunk` provides, such as ChunkPos and BlockID. This commit makes use of Cargo's "workspaces" feature, which was made for this type of scenario.

Also:
* It's called `chunk` now, as it does a lot more than just generate them.
* build.py is much improved.